Refactor createUser and add setAvailability
