+++ src/thunks/createUser.js
@@ -1,8 +1,7 @@
 import { fetchData } from '../utils';
 import { setError, setLoading, setUser } from '../actions';
-import { getSchedule } from './getSchedule';
 import * as gql from '../queries';
-import * as helpers from '../helpers';
+import { setAvailability } from './setAvailability';
 
 export const createUser = (userInfo, availabilities) => {
   return async dispatch => {
@@ -10,25 +9,13 @@ export const createUser = (userInfo, availabilities) => {
     try {
       const userQuery = gql.createUser(userInfo);
       const { user } = await fetchData(userQuery);
-      const startDate = 'Mon Apr 29 2019';
-      const inning = helpers.getDatesToDisplayFancy(startDate, 29);
-      const unformattedPairings = helpers.createPairingsForQuery(
-        availabilities,
-        inning,
-        user.id
-      );
-      const formattedPairings = helpers.formatPairingsForQuery(unformattedPairings)
-      const pairingQuery = gql.createPairings(formattedPairings);
-      await fetchData(pairingQuery);
+      if (availabilities) {
+        dispatch(setAvailability(user.id, availabilities));
+      }
       dispatch(setUser(user));
-      dispatch(getSchedule(user.id));
     } catch (error) {
       dispatch(setError(error.message));
     }
     dispatch(setLoading(false));
   };
 };
-
-
-
-
